Crate oak_config_serializer

An utility for serializing an Oak application configuration.

Invoke with:

cargo run --manifest-path=oak/server/rust/oak_config_serializer/Cargo.toml -- \
    --input-file=<CONFIGURATION_FILE> \
    --output-file=<OUTPUT_FILE>

Structs

Config

External

InitialNodeConfig

Opt
Command line options for the Oak Application Configuration Serializer.

Enums

Module

Constants

CACHE_DIRECTORY
Directory used to save downloaded Wasm modules. Created in the std::env::current_dir().

Functions

download_file_from_url
Download file from url.

get_module_cache_path
Get path for caching a downloaded file in the CACHE_DIRECTORY. Cache file is named after sha256_sum.

get_sha256
Computes SHA256 sum from data and returns it as a HEX encoded string.

load_module
Load Wasm module from file or URL if specified. If the file was downloaded from URL, it is cached in CACHE_DIRECTORY.

main

write_config_to_file
Serializes an application configuration from app_config and writes it into filename.